+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - `read` now supports datetimes, with new keyword arguments `convert_dates`,
   `keep_default_dates` and `date_unit`.
+- `write` now supports datetimes, with new keyword arguments `date_format` and `date_unit`.
 
 ## 1.0.0 (2023-02-07)
 - Initial release.